chore(runners.go): extract cache related elements
